+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 - Supports GFA1 and subset of GFA2
 - Supports a variety of force based layout settings
+- Saves state in URL params